Working on feedback with connected components and laser distinguishing.

+++ include/image_processing.hpp
@@ -3,7 +3,6 @@
 #include <cuda_runtime.h>
 #include <opencv2/opencv.hpp>
 #include "frame.hpp"
-#include "two_pass_algorithm.hpp"
 
 namespace gpu {
 
@@ -29,7 +28,7 @@ void closing();
 void open_and_close();
 void close_and_open();
 
-std::pair<uint16_t, uint16_t> detect_laser(uint8_t* red_frame,
+std::pair<int32_t, int32_t> detect_laser(uint8_t* red_frame,
                                            uint8_t threshold);
 
 }  // namespace gpu
